# documentation/backend/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alforis CRM API",
    "version": "1.0.0",
    "description": "Documentation de référence de l'API CRM Alforis. Les endpoints décrits couvrent les fonctionnalités principales livrées dans le plan d'amélioration : authentification, organisations, personnes, tâches, mandats, recherche globale, exports et webhooks.",
    "contact": {
      "name": "Équipe CRM Alforis",
      "email": "tech@alforis-finance.com"
    }
  },
  "servers": [
    {
      "url": "https://crm.alforis-finance.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Développement local"
    }
  ],
...
  "paths": {
...
    "/api/v1/mandats": {
      "get": {
        "tags": [
          "mandats"
        ],
        "summary": "Lister les mandats",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "organisation_id",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "proposé",
                "signé",
                "actif",
                "terminé"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page de mandats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMandats"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "mandats"
        ],
        "summary": "Créer un mandat",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MandatCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mandat créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mandat"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/mandats/active": {
      "get": {
        "tags": [
          "mandats"
        ],
        "summary": "Mandats actifs",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "organisation_id",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mandat"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/mandats/organisation/{organisation_id}": {
      "get": {
        "tags": [
          "mandats"
        ],
        "summary": "Mandats d'une organisation",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "organisation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mandat"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/mandats/{mandat_id}": {
      "get": {
        "tags": [
          "mandats"
        ],
        "summary": "Détail d'un mandat",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mandat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mandat détaillé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MandatDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "mandats"
        ],
        "summary": "Mettre à jour un mandat",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mandat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MandatUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mandat mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mandat"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "mandats"
        ],
        "summary": "Supprimer un mandat",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mandat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Supprimé"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/search/mandats": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Recherche mandats",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Résultats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Organisation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "ACME Capital"
          },
          "category": {
            "type": "string",
            "enum": [
              "Institution",
              "Wholesale",
              "SDG",
              "CGPI",
              "Autres"
            ]
          },
          "aum": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "aum_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "strategies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "website": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "country_code": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "nullable": true
          },
          "domicile": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "minLength": 2,
            "maxLength": 5,
            "example": "FR"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "language",
          "is_active"
        ]
      },
...
      "Mandat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "organisation_id": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "proposé",
              "signé",
              "actif",
              "terminé"
            ]
          },
          "date_signature": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "date_debut": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "date_fin": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "is_actif": {
            "type": "boolean",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "organisation_id",
          "status"
        ]
      },
      "MandatCreate": {
        "type": "object",
        "properties": {
          "organisation_id": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "proposé",
              "signé",
              "actif",
              "terminé"
            ],
            "default": "proposé"
          },
          "date_signature": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "date_debut": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "date_fin": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "organisation_id"
        ]
      },
      "MandatUpdate": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "proposé",
              "signé",
              "actif",
              "terminé"
            ],
            "nullable": true
          },
          "date_signature": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "date_debut": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "date_fin": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "MandatDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Mandat"
          },
          {
            "type": "object",
            "properties": {
              "organisation": {
                "$ref": "#/components/schemas/Organisation"
              },
              "produits": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "isin": {
                      "type": "string",
                      "nullable": true
                    },
                    "type": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "type",
                    "status"
                  ]
                }
              }
            },
            "required": [
              "organisation"
            ]
          }
        ]
      },
      "PaginatedMandats": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "skip": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Mandat"
            }
          }
        },
        "required": [
          "total",
          "skip",
          "limit",
          "items"
        ]
      },
...
      "SearchCollection": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "label": {
                  "type": "string"
                },
                "excerpt": {
                  "type": "string",
                  "nullable": true
                },
                "score": {
                  "type": "number",
                  "format": "float",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "label"
              ]
            }
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items",
          "limit",
          "offset"
        ]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Auth requise",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Erreur de validation",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "loc": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "msg": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Ressource introuvable",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}